Performance improvements
* Use cmov instruction on amd64 for constant time move
* FeZero and FeOne unrolled

benchmark                    old ns/op     new ns/op     delta
BenchmarkKeyGeneration-4     64769         55550         -14.23%
BenchmarkSigning-4           65107         56922         -12.57%
BenchmarkVerification-4      160966        161068        +0.06%
